Refactored code and tests for TM3 message encoding/decoding
